Pass rid == -1 to iflib_softirq_alloc_generic() for bnxt
bnxt doesn't use shared interrupts, so there should be no need to
try to get TX queues on corresponding RX interrupt CPUs.  This
sbould be benchmarked and verified.  If it does need to set affinity,
look at how ixgbe and e1000 do it since the rid should not be the
qidx anymore.